backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
openai>=1.50.0
python-dotenv==1.0.1
ddgs>=6.0.0
